requirements.txt
redis==3.0.1
psycopg2-binary==2.7.6.1
pika==0.13.0
Flask==1.0.2
celery==4.2.1
Telethon==1.5.4
Flask-Gravatar==0.5.0
